requirements.txt
torch==1.6.0
imageio==2.9.0
numpy==1.20.2
kivymd==0.104.2
torchvision==0.6.0
certifi==2020.12.5
torch_geometric==1.7.0
h5py==3.2.1
Pyrebase4==4.5.0
ortools==8.2.8710
opencv_python==4.5.2.52
moviepy==1.0.3
Kivy==2.0.0
matplotlib==3.4.1
requests==2.25.1
evaluate==0.0.3
Pillow==8.3.2
pyrebase==3.0.27
PyYAML==5.4.1
scikit_learn==0.24.2
